+++ src/com/yanncebron/m68kplugin/structureview/M68kStructureViewModel.java
@@ -74,7 +74,8 @@ public class M68kStructureViewModel extends StructureViewModelBase implements St
         M68kBundle.message("structureview.filter.macros"), M68kIcons.LABEL_MACRO, "SHOW_MACROS"),
       createFilter(createLabelFilteringFunction(
         EnumSet.of(M68kLabelBase.LabelKind.EQU, M68kLabelBase.LabelKind.EQUALS,
-          M68kLabelBase.LabelKind.SET, M68kLabelBase.LabelKind.EQUR)),
+          M68kLabelBase.LabelKind.SET, M68kLabelBase.LabelKind.EQUR,
+          M68kLabelBase.LabelKind.FO)),
         M68kBundle.message("structureview.filter.assignments"), M68kIcons.LABEL_EQU, "SHOW_ASSIGNMENTS")
     };
   }